+++ Content.Client/Options/UI/Tabs/MiscTab.xaml.cs
@@ -52,7 +52,7 @@ public sealed partial class MiscTab : Control
         Control.AddOptionCheckBox(CCVars.ChatEnableFancyBubbles, FancySpeechBubblesCheckBox);
         Control.AddOptionCheckBox(CCVars.ChatFancyNameBackground, FancyNameBackgroundsCheckBox);
         Control.AddOptionCheckBox(CCVars.StaticStorageUI, StaticStorageUI);
-        Control.AddOptionCheckBox(CCVars.OfferModeIndicatorsPointShow, ShowOfferModeIndicatorsCheckBox);
+        Control.AddOptionCheckBox(Shared.Backmen.CCVar.CCVars.OfferModeIndicatorsPointShow, ShowOfferModeIndicatorsCheckBox);
         Control.Initialize();
     }
 }
