+++ Content.Shared/Backmen/Alert/Click/AcceptingOffer.cs
@@ -1,0 +1,8 @@
+using Content.Shared.Alert;
+
+namespace Content.Shared.Backmen.Alert.Click;
+
+/// <summary>
+/// Accepting the offer and receive item
+/// </summary>
+public sealed partial class AcceptOfferAlertEvent : BaseAlertEvent;
