+++ Content.Shared/Backmen/CCVar/CCVars.cs
@@ -184,5 +184,7 @@ public sealed class CCVars
 
     public static readonly CVarDef<bool> AutoGetUp =
         CVarDef.Create("laying.auto_get_up", true, CVar.CLIENT | CVar.ARCHIVE | CVar.REPLICATED);
-
+    
+    public static readonly CVarDef<bool> OfferModeIndicatorsPointShow =
+        CVarDef.Create("hud.offer_mode_indicators_point_show", true, CVar.ARCHIVE | CVar.CLIENTONLY);
 }
